Agrego funcionalidad a productoServicesImpl

+++ src/main/java/com/tienda/entity/Producto.java
@@ -63,7 +63,7 @@ public class Producto {
         return cantidad;
     }
 
-    public void setCantidad(int cantidad) {
+    public void setCantidad(double cantidad) {
         this.cantidad = cantidad;
     }
